created a new shardstrategy package
